Refactorizando clase FileRepository
